requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
numpy>=2.0.0
